Разделить код на несколько файлов
