+++ bestchange_api/__init__.py
@@ -1,1 +1,1 @@
-from .main import BestChange
+from .bestchange import BestChange
